+++ packages/observable-client/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Error when unable to track a transaction because we haven't fully recovered from a `stop` event
+- Fix `readyBlocks$` stream
 
 ### Added